# Experiment with using hash table from COVER 