Changes Made:
- Originally, an array called segmentFreqs of size 2^f is used to keep track of number of occurrence of a dmer in the current segment. In this pull request, the hash map from COVER is used to keep track of this information.

Performance:
- Time significantly increases since it requires a lot of time to group the dmers and assign dmerid, which is what makes COVER slow. 
- Compression ratio increases sometimes and decreases sometimes. 
- Based on analysis from Instruments, time spent on ctx_init went from 66ms to 8.48s for optimize FASTCOVER since sorting takes a large portion of time. Time spent on buildDictionary, where selectSegment takes a large portion of time, went from 4.41s to 8.33s. 
- Since the compression ratio doesn't change much from this change but the speed got a lot slower, it doesn't seem like this change should be merged.