@terrelln 2018-08-07
Cool, looks like this is slightly slower still, even with `f=23`, lets stick with the `segmentFreqs` table.